Fix compiler warnings in camel.

+++ servers/exchange/storage/e-folder-exchange.h
@@ -55,7 +55,7 @@ const gchar *e_folder_exchange_get_permanent_uri    (EFolder    *folder);
 void        e_folder_exchange_set_permanent_uri    (EFolder    *folder,
 						    const gchar *permanent_uri);
 
-long long int	e_folder_exchange_get_folder_size (EFolder *folder);
+gint64	e_folder_exchange_get_folder_size (EFolder *folder);
 void		e_folder_exchange_set_folder_size (EFolder *folder, gint64 folder_size);
 
 gboolean    e_folder_exchange_get_has_subfolders   (EFolder    *folder);
